requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1

# Slack
slack-bolt==1.21.2
slack-sdk==3.33.4

# AI/LLM
google-generativeai==0.8.3
openai==1.58.1
anthropic==0.40.0

# Database
asyncpg==0.30.0
redis==5.2.1

# HTTP Client
httpx==0.28.1
aiofiles==24.1.0

# Document Processing
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3

# Utilities
pydantic==2.10.3
pydantic-settings==2.6.1
tenacity==9.0.0

# Monitoring
structlog==24.4.0
prometheus-client==0.21.1
